+++ lib/keen_pure_admin/components/kpi_bento.ex
@@ -151,15 +151,19 @@ defmodule PureAdmin.Components.KpiBento do
         </div>
       <% end %>
 
-      <Kpi.kpi_detail
-        :if={@has_detail?}
-        title_text={@detail_title_text}
-        rows={@detail_rows || auto_rows(assigns)}
-      >
-        <%= for d <- @detail do %>
-          {render_slot(d)}
-        <% end %>
-      </Kpi.kpi_detail>
+      <%= if @detail != [] do %>
+        <Kpi.kpi_detail :if={@has_detail?}>
+          <%= for d <- @detail do %>
+            {render_slot(d)}
+          <% end %>
+        </Kpi.kpi_detail>
+      <% else %>
+        <Kpi.kpi_detail
+          :if={@has_detail?}
+          title_text={@detail_title_text}
+          rows={@detail_rows || auto_rows(assigns)}
+        />
+      <% end %>
     </div>
     """
   end